requirements.txt
pandas
pyodbc
python-dotenv
sqlalchemy
langchain
langchain-openai
langchain-community
faiss-cpu
